pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-issue-tracker"
version = "0.1.0"
description = "CSV-first vLLM issue snapshot pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.0",
    "tqdm>=4.60",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

